+++ src/main/java/pe/edu/upc/abilityhelpv1/servicesimplements/UserServiceImplement.java
@@ -2,11 +2,19 @@ package pe.edu.upc.abilityhelpv1.servicesimplements;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import pe.edu.upc.abilityhelpv1.dtos.UserDTO;
+import pe.edu.upc.abilityhelpv1.entities.Gender;
+import pe.edu.upc.abilityhelpv1.entities.Personalities;
+import pe.edu.upc.abilityhelpv1.entities.Rol;
 import pe.edu.upc.abilityhelpv1.entities.User;
+import pe.edu.upc.abilityhelpv1.repositories.IGenderRepository;
+import pe.edu.upc.abilityhelpv1.repositories.IRolRepository;
 import pe.edu.upc.abilityhelpv1.repositories.IUserRepository;
+import pe.edu.upc.abilityhelpv1.servicesinterfaces.IPersonalityServices;
 import pe.edu.upc.abilityhelpv1.servicesinterfaces.IUserServices;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class UserServiceImplement implements IUserServices {
@@ -22,17 +30,18 @@ public class UserServiceImplement implements IUserServices {
         return sR.findAll();
     }
     @Override
-    public void delete(int id) {
+    public void delete(Long id) {
         sR.deleteById(id);
     }
     @Override
-    public User listarId(int idUsuario) {
+    public User listarId(Long idUsuario) {
         return sR.findById(idUsuario).orElse(new User());
     }
     @Override
     public void update(User user) {
         sR.save(user);
     }
+
 
 
 }
